+++ src/Frontend/Utils/constants.ts
@@ -22,7 +22,7 @@ const apiUrl = 'https://9a46-143-244-168-87.ngrok.io/subgraphs/name/df-arena-v2'
 
 const competitiveConfig = '0x568297442f966cc66f2be7ced683e35ea2ca1e68b4f26dd5424158244da40bcc';
 
-const tutorialConfig = '0x0000000000000000000000000000000000000000000000000000000000000001';
+const tutorialConfig = '0xb03e4e860ff6f0351b5ef6d04061f3b9dddd81e1e77ce1a74f1c535b7aaa5224';
 
 const roundStartTimestamp = '2022-07-13T00:00:00.000Z';
